Use generic cell
