+++ Demo/Sources/ShowcaseView.swift
@@ -6,26 +6,6 @@
 
 import SwiftUI
 
-// Behavior: h-hug, v-hug
-private struct Cell<Presented: View>: View {
-    let title: String
-
-    @ViewBuilder var presented: () -> Presented
-    @State private var isPresented = false
-
-    var body: some View {
-        Button(action: action) {
-            Text(title)
-                .foregroundColor(.primary)
-        }
-        .sheet(isPresented: $isPresented, content: presented)
-    }
-
-    private func action() {
-        isPresented.toggle()
-    }
-}
-
 // Behavior: h-exp, v-exp
 struct ShowcaseView: View {
     var body: some View {
